Implement peephole optimizations for value classes
The extractors in TreeExtractors are copied or inspired from
src/compiler/scala/tools/nsc/ast/TreeInfo.scala in scalac.

Note that there is no special rule for the following optimization for
value classes from SIP-15:
  new V(e) == new V(f) => e == f
  new V(e) != new V(f) => e != f
This is because PeepholeOptimize is after InterceptedMethods and
InterceptedMethods may transform ==/!=. We would have to duplicate this
logic in PeepholeOptimize if we had these rules. Instead,
ExtensionMethods#transformApply does:
  v1 == v2 => v1.u == v2.u
  v1 != v2 => v1.u != v2.u
Then PeepholeOptimize applies the general transformation:
  (new V(e)).u => e
This is equivalent to the original rules from SIP-15 but means that we
do not have to handle calls to ==/!= after InterceptedMethods