NavigateAST#pathTo: prefer non-zero-extent position
When determining the tree corresponding to a position, if there was an
ambiguity between two trees because they both contained the position, we
used to always choose the first one. Instead, it's more useful to prefer
trees where pos.start != pos.end because they're more likely to
correspond to something the user wrote, this is useful in IDEs.