Better handling of Contexts in RunInfo and Implicits
In several situations, instead of the common pattern in the compiler of
using the enclosing Context available coming an implicit method
parameter, we used some other Context. So far, this wasn't an issue but
starting with the next commit it is important for FunProto#typedArgs to
be called in a Context whose owner chain contains the Context where the
FunProto was created. This motivated the following changes:

- RunInfo does not actually need to keep a reference to the root Context:
  It was previously used by `ImplicitRunInfo#implicitScope`, but can be
  replaced by using the Context from the enclosing scope (which was already
  used in this method with the name `liftingCtx`).
- OfTypeImplicits does not need to keep a reference to the root Context
  either, we can just replace its lazy vals by defs that cache their results.
- ContextualImplicits does need to keep track of the Context from which
  its implicit references come from, but this Context itself does not have
  to be available implicitly in the class, instead the Context from the
  enclosing scope is used.

This also requires replacing OfTypeImplicits#toString and
ContextualImplicits#toString by a toText method in Printer to get access
to a Context.