Fix instantiatability checking
We used to check only New nodes for instantiability. But sometimes
these nodes have type constructors as types. The previous implementations
of generics hid the problem, but in the new scheme the subtype test
between New's type and its given self type fails if New's type lacks
arguments.

To compensate, we need to check the constructor call instead of the
New itself.