Ensure type correctness of repeated arguments
Make sure elimRepeated produces array arguments that match the formal
parameter type. Previously, there was a gap which showed itself when compiling

    scala/concurrent/forkjoin/package.scala

There, the method expected a repeated argument of type ForkJoinTask[_], so the
expected type was Array[ForkJoinTask[_]]. But the type computed by seqToArray
was Array[ForkJoinTask[T]]. Since arrays are non-variant, these types are not
compatible! This somehow slipped through with the previous implementation of
applied types.