Partially fix null unboxing in lambdas
When unboxing a null to a value class in Scala, the default value of the
value class should be returned, before this commit an NPE was thrown instead.
Like in https://github.com/scala/scala-java8-compat/commit/d7e0125fc945b855cc9cd05489985b6f689d551c,
we fix this by using BoxesRunTime, bun we don't use it for boxing
since it's unnecessary.

After this commit, we still throw NPE when unboxing nulls in
unspecialized lambdas called in generic contexts (see if3_unspecialized
in the added test). This is addressed by the next commits in this PR.