Allow for aut-updating of package members
Previously, an overwritten package member always got the denotation of the
new package, so package members always auto-updated, and we could not even
detect that they were no longer valid. This is a problem because it means that
TermRefs and TypeRefs pointing to overwritten members will not be updated.

We now do something less sweeping. A call to `current` will invoke `bringForward`
which will do the update if the `autoUpdatePackageMembers` config flag is set.
After the call to `current` the old symbol will be considered valid, which is a
potential problems because we then have two valid symbols pointing to the same
denotations. So it would be best if we could set `autoUpdatePackageMembers = false`.