Fix #3608: Treat outer selections as separate Tasty tree
Tasty needs to know about outer selections as separate trees. So instead of
exposing a name for outer selects we expose a tree format. This allows us to
store the type of tree selection in the tree, which got lost before.

Also: Change format of NAMEDarg to no longer require a length.