Adapt/remove tests for the 2.13 stdlib.
For the tests that were originally copied from scalac, I also copied the
updated version if one existed, or removed it if they had removed it.

Unfortunately, I also had to remove a bunch of Dotty tests that relied
on details of the 2.12 stdlib that I couldn't easily port to the 2.13
stdlib.

i3518 has to be disabled because using scala.Unit is marked @compileTimeOnly
and dotty crashes on @compileTimeOnly right now.