Revert "variance 0 for prefix, -1 doesn't seem to make sense and doesn't match comment"
This reverts commit 4a692d76a3960b5632292723c49c9bcd384c1c86.
This reverts commit 2ccef60145297966e744bdad2a235728edc23ba5.

I misread. Fixes, tests/run/virtpatmat_stringinterp.scala.

A better way to think about it is probably: it's safe to use
range(Nothing, ...) because the prefix is never contravariant, it
wouldn't be safe to use range(..., Any) however.
Better solution: do not limit variance, but when prefix doesn't have
member, widen. More powerful in principle, but can we find a testcase?