sbt-dotty: Rework classpath and ScalaInstance handling
In particular, stop relying on zinc handling of the ScalaTool
configuration (by setting managedScalaInstance to false) and instead
fetch artifacts ourselves. Also document clearly what should end up on
which classpath.

Also cleanup the handling of ScalaInstance in the build of Dotty itself,
this will be more important once we switch to Dotty as our stage0 compiler.