Fix #6460: less forcing in checkNonCyclic
checkNonCyclic was forcing too many infos because it didn't account for
the case where the prefix is a ModuleVal and not a ModuleClass. Fixing
this solves the incremental compilation issue in the Dotty build
reported in #6460 were the root cause was a completion cycle when
jointly compiling SymbolLoaders and Types that looked like this:

Complete `object tpd`
 Complete its parent `Trees.Instance`
  Complete its type parameter `T >: Untyped <: Type`
    run `checkNonCyclic` on the the type parameter definition
      Complete `Type`
        Complete `import ast.tpd._` in `Types.scala`
          Eventually this requires unpickling the selection `Trees.Instance#T` which
          gets a `NoDenotation` since `T` hasn't been entered yet.

As a side-effect this also means that the neg/toplevel-cyclic
test no longer leads to an "illegal cyclic reference" error, instead it
now causes "recursion limit exceeded" errors, which are less nice but
still acceptable.