Add new SymbolLoader: NoLoader
When a SymbolLoader gets completed, we sometimes use NoCompleter as an
intermediate info, we now use NoLoader instead. This will be useful in
the next commit where we restrict which completers are allowed to set
`isAbsent`. We must allow SymbolLoader to set `isAbsent` , but we don't
want to allow any NoCompleter to set it, so instead we use NoLoader
which is a subtype of both SymbolLoader and NoCompleter.

It also turns out that even without the changes to `isAbsent`, using a
`SymbolLoader` here is a good idea since `SymDenotation#isCurrent`
handles SymbolLoader specially and did not take into account that they
were sometimes replaced by NoCompleter. This means that checking for the
presence of a flag can now lead to more completion, which required
adding an early exit in `Types#isArgPrefixOf` to avoid a cyclic
reference.