Add new SymbolLoader: NoLoader
When a SymbolLoader gets completed, we sometimes use NoCompleter as an
intermediate info, we now use NoLoader instead. This will be useful in
the next commit where we restrict which completers are allowed to set
`privateWithin`. We must allow SymbolLoader to set `privateWithin`
because its value cannot be known at symbol creation time, but we don't
want to allow any NoCompleter to set it, so instead we use NoLoader
which is a subclass of SymbolLoader.