Drop annotations from constructor tparams
Scala 2 preserves annotations both on the param accessors and on
compiler-generated methods such as copy, so we should do the same. But
there doesn't seem to be much point in preserving them on the type
parameters of the constructors. Removing them avoids a cyclic reference
error involving `@specialized` which happened in the past and resurfaced
with the completions-related changes in this PR.