Add explicit check against self-annotated classes
Following the completions-related changes in this PR, we no longer
encounter cyclic reference errors when annotating a class with
itself (i1212.scala) as well as in some situation where an annotation
class indirectly refers to itself (i3506.scala).

While the latter is OK, the former is problematic because when
unpickling from tasty, the modifiers and annotations of a symbol are read
before the symbol itself is created, so we can't support self-annotation
classes, we therefore need to disallow them explicitly.