Properly unpickle Scala 2 refinements
They're pickled with a class symbol, but we unpickled them as a type
aliases, which meant that the logic in REFINEDtpe wasn't able to
substitute `this` references correctly (there was even a comment
expressing puzzlement about this). Instead, unpickle them as
classes (TODO: add some logic in TreeChecker to make sure these symbols
are never referenced anywhere).