Fix typing of PartialFunction argument to overloaded method
Applications#preTypeArgs can type arguments with an expected type like
`PartialFunction[Int, WildcardType]`, before this commit this was
rejected by `typedClosure` because it requires the expected type to be
fully-defined. Fixed by replacing the underdefined expected type by a
new one where the argument types come from the closure method type.

This is useful in particular to type `Map#collect` in 2.13 which is
overloaded.