Fix flag mutation for Scala2xPartiallyAugmented
Calling `setFlag` on the current denotation is not enough: there might
be other more recent denotations created, if we don't also set the flag
on these denotations, then the `is(Scala2xPartiallyAugmented)` check in
`LinkScala2Impls` will fail.

FIXME: we probably need to do the same for all the other usages of
setFlag/resetFlag/flags_= in the compiler.