Fix method calls on a Java annotation
In the compiler we pretend that Java annotations are classes, but
they're really interfaces from the JVM point of view, so we need to
treat them as such in the backend when calling a method on them,
otherwise we'll emit invalid bytecode.

Needed to get tests/run/i2760 to pass after the previous commit.