Do not type annotations twice
0500452f3ae23a7da55e42bbfc6d36117a041262 replaced the usage of
`typedAheadAnnotation` by `typedAnnotation` in
`Namer#completeAnnotations`, but it left the call to `typedAnnotation`
in `Typer` which means annotations were typed twice.

Also, make sure that annotations are typed in the correct context
derived from `annotContext`, previously this was the case when they were
typed a second time in Typer but not when they were typed in Namer, this
required a change in ReifyQuotes which was so far compensating for the
previous incorrect behavior.