Take expected type into account when typing a sequence argument
Given an expected type `*T`, we allow a sequence argument `xs: _*` to be
either a `Seq[T]` or an `Array[_ <: T]`, irrespective of whether the
method we're calling is a Java or Scala method. So far we typed sequence
arguments without an expected type, meaning that adaptation did not take
place. But thanks to #8635, we can type them with an expected type of
`Seq[T] | Array[_ <: T]` and type inference works out. This is what this
commit does.