Fix ClassCastException with Java primitive varargs
The class constant corresponding to the primitive type `int` is
represented in bytecode by `Integer.TYPE`, this is correctly handled
when calling `clsOf`, but so far `ElimRepeated` was manually creating a
constant instead which does not work (I've now changed the backend to
emit a failure instead of doing the wrong thing when someone does that,
I also plan to rework the way we handle classOf to make this less
error-prone).

Note: after this commit, tests/run/t1360.scala started failing, this is
fixed in the next commit.