Constraint#approximation: Fix unsound approximation
`approximation` used to replace recursive occurences of the type
variable being instantiated by `TypeBounds.empty`, but this is not sound
in general since the approximation might be outside the bounds of the
type variable. We could try to do something fancy with
ApproximatingTypeMap but it seems complicated (and expensive) and
there's an easier solution: since last commit, only the upper bound is
allowed to be recursive, so we can just instantiate the type variable to
its lower bound to avoid the problem in a sound and cheap way.