Drop filler from name encoding to keep signatures stable
Signatures should not change before erasure, but the ByNameClosure phase
changes the owner of some definitions, which can change their full name
and therefore any signature which refers to these names.

This commit fixes this by dropping the "_$" fillers added in fullName
which doesn't seem necessary and lead to unstable names. The check for
signature consistency in TreeChecker was also improved as it wasn't able
to catch this issue before due to signature caching hiding the issue.