ClassfileParser: Parse varargs in a more direct manner
Varargs are encoded as arrays in methods with a special flag, previously
we parsed them as arrays then translated the array into a repeated
param. With this commit we directly parse them as array, this isn't
really any simpler since we need to keep track of the fact that we're in
a varargs method, but it will become useful in the next commit where we
start treating arrays differently from varargs (`T[]` will still be
translated as `Array[T & Object]` whereas `T...` will be translated as
`T*` and handled specially in ElimRepeated to preserve soundness).