Avoid forcing nested annotation
Java allows annotations inside arrays inside annotations, when parsing
the outer annotation, we need to produce an untyped tree for the array
and therefore the nested annotations. Previously this was done using
`Annotation#tree` which actually produces a typed tree, this works
since typed trees can be used as untyped tree but it requires forcing
the content of the nested annotation which can lead to cycle. Instead,
we now use a dedicated subclass of Annotation with an extra `untpdTree`
method we use for the purpose of generating a tree for the nested
annotation without forcing anything.