Fix overcompilation when inheriting inline defs
When the body of an inline def changes, all its callers need to be
recompiled. So when producing an API signature for an inline def to be
sent to Zinc, we need to somehow include that body. So far, this was
done by simply including the `toString` of the tree in the API
signature, but the API of a class includes its inherited members, and
when calling `toString` on an inherited inline def, its body ends up
being printed as just `TreeUnpickler$LazyReader@123...` which is clearly
not what we want, in practice this lead to overcompilation rather than
undercompilation because the string includes the identityHashCode of the
LazyReader which changes at each compilation run.

This commit fixes this by implementing a proper way to hash trees, this
is still not perfect since types aren't included in the hash (see
comments in the code), but that can be addressed at a later point (and
the exact strategy we want to use also depends on how we evolve Zinc).