Add a way to run specific tests on the CI
With this commit, the CI will look for some magic strings in commit
messages to decide what tests to run, for example if
"[test_non_bootstrapped]" anywhere in the commit message of a PR, the
non-bootstrapped tests will be run on top of the normal tests we run for
PRs. It's also possible for example to turn off the windows tests with
"[skip test_windows]".