Get rid of JavaMethodType
This is possible thanks to the previous commit, in all remaining places
where we called `isJavaMethod` we can rely on contextual informations to
obtain the same information, so we can completely get rid of it which is
a nice simplification.

Also get rid of ElimRepeated#transformTypeOfTree which turned out to be
a no-op: the denotation transformer already transforms method types and
there's nothing to do on top of that.