Replace SingleDenotation by SymDenotation after erasure
This is an alternative fix for #11303, the downside is that when
time-travelling from post-erasure to pre-erasure, we need to recompute
the SingleDenotation. This commit passes all tests except
tests/pos/t2399.scala which fails compilation from tasty. This could
probably be fixed by changing the logic in computeDenot to only run
`memberDenot` for opaque types, for other infos we can just use
`asSeenFrom` instead.