Add MethodOrPoly#signature(isJava) overload
This makes it possible to compute the Java or Scala signature of a
method, regardless of whether it was defined as a JavaMethod or not.
Because there's nothing left that is shareable, this required removing
SignatureCachingType and duplicating the signature caching logic between
MethodOrPoly and NamedType.

This commit itself does not change any semantics, but this new method
will be useful for the next commit.