Only use Java signatures for SymDenotations
For SymDenotations, it's important that we use Java signatures for Java
definitions to maintain the invariant that two overloads in the same
owner have different signatures, but for non-SymDenotations we don't
make this guarantee so we could go either way.

I originally chose to treat them like SymDenotation for consistency, but
this is actually a bit weird: because these denotation infos are seen
from a prefix that can be Scala-defined, it means that we can end up
erasing union types in a Java signature. This will become more
problematic when we add Scala 2 signatures later in this PR because
Scala 2 union types should only appear in Scala.js code, and because the
Scala 2 intersection type erasure logic will be more complex and
therefore probably slower than the Scala 3 logic, so we should avoid
calling it when possible.

We sidestep all of this by ensuring that non-Scala-3 erasure is only
used for SymDenotations (whose info will never refer to Scala-3-specific
types).