SAM types cannot be sealed or final
Previously we caught this in Typer, but doing it in the `SAMType`
extractor is more robust because it is used in overloading resolution,
and this will be needed after the next commit to avoid selecting the
wrong overload in some cases.

This required dropping the final flag from the compiler-generated
ContextFunctionN, instead we explicitly prevent subclasses in RefChecks.