Fix subtype checking of instantiated TypeParamRef
If the TypeParamRef is already instantiated, we should recurse on its
instantiation. This was missing before this commit but was usually not a
problem since we normally reach TypeParamRef through TypeVar#underlying
which directly returns the instantiation if it exists. However, when
typing a method argument we check if its a subtype of the corresponding
formal parameter type which can be a TypeParamRef, and which might have
been instantiated when a previous argument was typed.