Fix nesting level to account for block scopes
Previously, the nesting level of a symbol was was always equal to the
nesting level of its owner incremented by one, but in a situation like
i8900a.scala where we have:

...
  val a = unwrap({
    class Local
...

the owner of both the type variable of `unwrap` and `Local` is `a`, and
so they ended up with the same level even though `Local` should not leak
outside of the block that defines it (i8900a.scala compiled regardless
due to other issues resolved later in this PR).

We rectify this by associating each scope with a level which is always
greated than the level of the enclosing scope (repurposing the existing
Scope#nestingLevel method which wasn't used for anything), newly created
symbols then simply take the level of the current scope (this required
tweaking typedCase so that the pattern symbols were typed in the scope
were they end up being entered).

Also add a `-Yprint-level` option for debugging level-related issues.