Fix variance of approximating map in `addOneBound`
It turns out the variance was flipped: when adding a constraint
`P <: List[WildcardType(Int, Any)]`, we should over-constrain to `P <: List[Int]`,
but previously we would under-constrain to `P <: List[Any]` which would
allow us later to infer `P := List[String]` for example. However, this
logic needs to be flipped when inferring necessary constraints (this explains
why the previous behavior was seemingly correct).

No test case but this will end up being important in later commits of
this PR where we re-use the same map to do more approximations.